fix: resolve pre-existing safety and concurrency issues (#226)
Addresses pre-existing code quality issues identified during clang-tidy
review (CodeRabbit findings on PR #225):

Critical data corruption / logic bugs:
- sd_message: fix IPv4MulticastOption length written to wrong indices
  (data[2]/data[3] → data[0]/data[1])
- standard_profile: reject duplicate E2E counters to prevent replay
  attacks (header.counter == last_counter now returns INVALID_ARGUMENT)

Concurrency / data race fixes:
- freertos/memory: make pool_initialized std::atomic<bool> with proper
  memory ordering to fix broken double-checked locking pattern
- threadx/memory: same fix for pool_initialized and guard_created
- tp_manager: collect timed-out transfers under lock, invoke
  completion_callback_ after releasing transfers_mutex_
- sd_client: copy callbacks from subscriptions/pending_finds maps under
  their respective locks, invoke them after lock release

Type safety / narrowing fixes:
- tp_reassembler: widen TP offset from uint16_t to uint32_t (28-bit
  field × 16 can exceed 64 KB)
- serializer: default DeserializationResult error_ to NOT_INITIALIZED
  instead of SUCCESS to prevent get_value() on empty optional
- e2e_crc: add PTRDIFF_MAX overflow guard before size_t → ptrdiff_t
  cast in iterator arithmetic

Closes #226

Made-with: Cursor

+++ include/tp/tp_types.h
@@ -66,7 +66,7 @@ struct TpConfig {
  */
 struct TpSegmentHeader {
     uint32_t message_length{0};     // Total message length
-    uint16_t segment_offset{0};     // Offset of this segment in the message
+    uint32_t segment_offset{0};     // Offset of this segment in the message
     uint16_t segment_length{0};     // Length of this segment's payload
     uint8_t sequence_number{0};     // Sequence number for ordering
     TpMessageType message_type{TpMessageType::SINGLE_MESSAGE};  // Type of TP message
@@ -104,8 +104,8 @@ struct TpReassemblyBuffer {
         start_time = std::chrono::steady_clock::now();
     }
 
-    bool is_segment_received(uint16_t offset, uint16_t length) const;
-    void mark_segment_received(uint16_t offset, uint16_t length);
+    bool is_segment_received(uint32_t offset, uint32_t length) const;
+    void mark_segment_received(uint32_t offset, uint32_t length);
     bool is_complete() const;
     std::vector<uint8_t> get_complete_message() const;
 };
